Enable inherited error checking tests.
