+++ sdks/python/apache_beam/runners/portability/flink_runner_test.py
@@ -32,6 +32,7 @@ from apache_beam.options.pipeline_options import StandardOptions
 from apache_beam.runners.portability import portable_runner
 from apache_beam.runners.portability import portable_runner_test
 from apache_beam.testing.util import assert_that
+from apache_beam.testing.util import equal_to
 
 if __name__ == '__main__':
   # Run as
@@ -108,6 +109,19 @@ if __name__ == '__main__':
 
     def test_pardo_timers(self):
       raise unittest.SkipTest("BEAM-4681 - User timers not yet supported.")
+
+    def test_assert_that(self):
+      # We still want to make sure asserts fail, even if the message
+      # isn't right (BEAM-6019).
+      with self.assertRaises(Exception):
+        with self.create_pipeline() as p:
+          assert_that(p | beam.Create(['a', 'b']), equal_to(['a']))
+
+    def test_error_message_includes_stage(self):
+      raise unittest.SkipTest("BEAM-6019")
+
+    def test_error_traceback_includes_user_code(self):
+      raise unittest.SkipTest("BEAM-6019")
 
     # Inherits all other tests.
